requirements/local.txt
# Local development: base + dev dependencies.
# Install with: pip install -r requirements/local.txt
# Prefer: uv sync (installs base + dev from pyproject.toml)

-r base.txt

# Dev / test / lint
coverage==7.13.4
django-coverage-plugin==3.2.0
django-debug-toolbar==6.2.0
django-extensions==4.1
django-stubs[compatible-mypy]==5.2.9
djangorestframework-stubs==3.16.8
djlint==1.36.4
factory-boy==3.3.2
ipdb==0.13.13
mypy==1.19.1
pre-commit==4.5.1
pytest==9.0.2
pytest-django==4.12.0
pytest-sugar==1.1.1
ruff==0.15.4
sphinx==9.1.0
sphinx-autobuild==2025.8.25
werkzeug[watchdog]==3.1.6
